+++ common/mpostgres/pagination.go
@@ -1,4 +1,4 @@
-package domain
+package mpostgres
 
 // Pagination is a struct designed to encapsulate pagination response payload data.
 type Pagination struct {
