+++ components/ledger/internal/ports/ledger.go
@@ -1,10 +1,10 @@
 package ports
 
 import (
+	"github.com/LerianStudio/midaz/common/mpostgres"
 	"os"
 
 	"github.com/LerianStudio/midaz/common"
-	"github.com/LerianStudio/midaz/components/ledger/internal/domain"
 	"go.mongodb.org/mongo-driver/bson"
 
 	"github.com/LerianStudio/midaz/common/mlog"
@@ -73,7 +73,7 @@ func (handler *LedgerHandler) GetAllLedgers(c *fiber.Ctx) error {
 
 	headerParams := common.ValidateParameters(c.Queries())
 
-	pagination := domain.Pagination{
+	pagination := mpostgres.Pagination{
 		Limit: headerParams.Limit,
 		Page:  headerParams.Page,
 	}
